refactor: Fetch high contrast state via a cross platform hook
