+++ src/hooks/useHighContrastState.ts
@@ -1,0 +1,6 @@
+'use strict';
+
+// Not Implemented on React Native Core
+export default function useHighContrastState() {
+  return false;
+}
